Remove ghc-lib-parser-ext dependency
Instead, hardcod most of its content in PyF.Internal.ParserEx.

The rational behind this choice is that ghc-lib-parser-ext is a super
cool library, but it does depends on cabal flags which changes how it
depends on `ghc` lib or `ghc-lib-parser`. I'd like to depend on `ghc`
lib (see previous commits) in order to reduce the number of
dependencies.