Feat: use GHC native error reporting
Instead of manually generating the error message with line / caret / ...
and passing this string to TH `fail`, we use GHC native error reporting
functionality.

The benefits are:
- Less code (all the manual formatting code is removed)
- More in sync with other errors
- KILLER FEATURE: the error location is now correctly reported by external
  tools, such as haskell-language-server. No more error spanning over
  your complete template haskell splice, the correct error will be
  pinpointed accurately.

This was achieved using an `unsafeCoerce`, to access the complete GHC
api. I've found about this hack when reading GHC source code and later
found that it was described in

https://www.tweag.io/blog/2021-01-07-haskell-dark-arts-part-i/